improve the security of RSA
In the past version, I made e as small as possible. This made x^e mod N
= x^e.
As we all know that, unicode ranges from 0000 to FFFF. If anyone got a
number like 1934917632, it's easy to realize that it's 72^5. So he/she
can chr(72) and get 'H', which is my RSA encrypted.